requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.2
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
legacy-cgi==2.6.3
loguru==0.7.3
packaging==24.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
schedule==1.2.2
setuptools==76.1.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
win32_setctime==1.2.0
